Switch module style to CommonJS
Did some more reading about JS modules and learned that the ES6 module syle
(which this codebase used previously) seems to be mostly unsupported and
experimental. So, for the time being, the exports style has been changed to
CommonJS. Since Node supports this normally, the testing process has been
simplified as well.